Refactor file structure so base classes are in cpo_base.py and OE/ELSFP functionality are in oe.py and elsfp.py
