+++ sonic_platform_base/sonic_xcvr/cpo/cpo_optoe_base.py
@@ -1,5 +1,6 @@
 from sonic_platform_base.sonic_xcvr.optoe_eeprom_rw import OptoeEepromReadWriteMixin
-from sonic_platform_base.sonic_xcvr.cpo.cpo_base import OeBase, ElsfpBase
+from sonic_platform_base.sonic_xcvr.cpo.oe import OeBase
+from sonic_platform_base.sonic_xcvr.cpo.elsfp import ElsfpBase
 
 class OptoeOeBase(OeBase, OptoeEepromReadWriteMixin):
     pass
