+++ tests/sonic_xcvr/test_elsfp.py
@@ -1,25 +1,13 @@
 from mock import MagicMock
 import pytest
 
-from sonic_platform_base.sonic_xcvr.cpo.cpo_api_factory import (
-    OeApiFactory,
-    ElsfpApiFactory,
-    CpoHardwareId,
-)
+from sonic_platform_base.sonic_xcvr.cpo.cpo_base import CpoHardwareId
+from sonic_platform_base.sonic_xcvr.cpo.elsfp import ElsfpApiFactory
 
 # OeId/ElsfpId currently define no members; use sentinels to represent
 # "an id is set" in tests that exercise the not-yet-supported code paths.
 SOME_OE_ID = object()
 SOME_ELSFP_ID = object()
-
-
-class TestOeApiFactory(object):
-    def test_create_api_raises(self):
-        oe = MagicMock()
-        oe.hardware_id = CpoHardwareId(oe_id=SOME_OE_ID, elsfp_id=None)
-        factory = OeApiFactory(oe)
-        with pytest.raises(ValueError):
-            factory.create_api()
 
 
 class TestElsfpApiFactory(object):
